Refactor timeout data types in coinflip program
- Updated the `max_timeout`, `min_timeout`, and `timeout` fields from `u16` to `u32` across multiple structs and functions to accommodate larger timeout values and improve consistency in data types.
- Adjusted related instructions and handlers to reflect these changes, enhancing the integrity and flexibility of game initialization and management.

These modifications improve the overall robustness of the coinflip program by refining timeout handling and ensuring appropriate data type usage.

+++ programs/coinflip/src/instructions/initialize_game.rs
@@ -9,7 +9,7 @@ pub fn handler(
     amount: u64,
     max_players: u16,
     min_players: u16,
-    timeout: u16,
+    timeout: u32,
     is_private: bool,
 ) -> Result<()> {
     // Initialize game
