+++ programs/coinflip/src/instructions/initialize_oracle.rs
@@ -5,8 +5,8 @@ pub fn handler(
     fee_percentage: u8,
     oracle_buffer_time: u16,
     max_players: u16,
-    max_timeout: u16,
-    min_timeout: u16,
+    max_timeout: u32,
+    min_timeout: u32,
 ) -> Result<()> {
     let oracle = &mut ctx.accounts.oracle;
     oracle.fee_percentage = fee_percentage;
